+++ CHANGELOG.md
@@ -13,6 +13,7 @@
   - 'filePath': new property for the file path.
   - 'fileOffset': new property for the offset inside filePath.
 - Fixed #162: Local stack not immediately update if value changed from UI.
+- Fixed #106: When loading a 128K .sna or .z80 file the port 7FFD is now set correctly (according file) for CSpect and dezogif (ZXNext).
 
 # 3.6.5
 - Debug command "-sjasmplus.path" added for dynamic loading of sld files during runtime (#156).