+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+# 3.7.1
+- #168: zx81 standard ula display now recognizes "jp (hl)" as end-of-screen
+
 # 3.7.0
 - New DZRP command CMD_INTERRUPT_ON_OFF in use for cspect. IMPORTANT: You need Cspect >= 3.0.15.2 to be compatible with DeZog.
 - Fixed issue loading .sna and .z80 files if interrupt is on and cspect is the target.